Backend implementation for course list sharing
This commit creates the schemas and endpoints to implement
course list sharing on the backend.

This commit creates one endpoint that supports two methods,
a GET method for fetching a course list and a POST method for
creating a course list. There is a schema for each method of
the endpoint to validate the input.

The GET method expects a uuid referring to the course list
in the query arguments of the request.

The POST method expects a course list in JSON format in the body
of the request.

This commit also contains code that references the list collection
in the Mongo database.