Frontend implementation for course list sharing
This commit implements the frontend components for
course list sharing.

This commit adds two routes to the frontend. The first
is the mylist route, this is the page where users can see
their local course list and publish it. The second is the
list route which displays published course lists.

To navigate to the local course list an option has been added
to the navbar.

To add/remove courses to the local course list an add/remove button
has been added to the course description page.

The local course list is stored in sessionStorage.